handle delete comment ketika ada child

+++ app/Http/Controllers/Api/CommentSubCommentController.php
@@ -7,7 +7,9 @@ use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\UpdateCommentRequest;
 use App\Http\Resources\PostCommentResource;
 use App\Models\Comment;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Http;
 
 class CommentSubCommentController extends Controller
 {
@@ -68,6 +70,7 @@ class CommentSubCommentController extends Controller
      */
     public function destroy(Comment $comment, Comment $subComment)
     {
+        if ($subComment->comments()->count() > 0) return response()->json(['message' => 'Comment cannot be deleted, someone comment at this already'], Response::HTTP_FORBIDDEN);
         $subComment->delete();
 
         return response()->json([
